apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "easylinkjobs-api"
version = "0.1.0"
description = "FastAPI backend for easylinkjobs"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.32.0,<1.0.0",
  "pydantic>=2.10.0,<3.0.0",
  "pydantic-settings>=2.7.0,<3.0.0",
  "email-validator>=2.2.0,<3.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "prisma>=0.15.0,<1.0.0",
  "redis>=5.2.0,<6.0.0",
  "PyJWT>=2.10.0,<3.0.0",
  "boto3>=1.36.0,<2.0.0",
  "cryptography>=44.0.0,<45.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "httpx>=0.28.0,<1.0.0",
  "black>=24.10.0,<25.0.0",
  "ruff>=0.8.0,<1.0.0",
  "mypy>=1.13.0,<2.0.0",
  "types-redis>=4.6.0.20241004"
]

[project.scripts]
seed-dev = "scripts.seed_dev:main"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
exclude = ["tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*", "scripts*"]
exclude = ["node_modules*", "prisma*", "tests*"]
